+++ lib/boruta/oauth/authorization/client.ex
@@ -35,7 +35,7 @@ defmodule Boruta.Oauth.Authorization.Client do
              }}
   def authorize(id: id, secret: secret, grant_type: grant_type)
       when not is_nil(id) and grant_type in ["revoke", "refresh_token"] do
-    with %Client{} = client <- ClientsAdapter.get_by(id: id),
+    with %Client{} = client <- ClientsAdapter.get_client(id),
          true <- Client.grant_type_supported?(client, grant_type) do
       case {apply(Client, :"public_#{grant_type}?", [client]),
             Client.check_secret(client, secret)} do
@@ -74,7 +74,7 @@ defmodule Boruta.Oauth.Authorization.Client do
 
   def authorize(id: id, secret: secret, grant_type: grant_type)
       when not is_nil(id) and not is_nil(secret) do
-    with %Client{} = client <- ClientsAdapter.get_by(id: id),
+    with %Client{} = client <- ClientsAdapter.get_client(id),
          :ok <- Client.check_secret(client, secret),
          true <- Client.grant_type_supported?(client, grant_type) do
       {:ok, client}
@@ -99,7 +99,7 @@ defmodule Boruta.Oauth.Authorization.Client do
 
   def authorize(id: id, redirect_uri: redirect_uri, grant_type: grant_type)
       when not is_nil(id) and not is_nil(redirect_uri) do
-    with %Client{} = client <- ClientsAdapter.get_by(id: id),
+    with %Client{} = client <- ClientsAdapter.get_client(id),
          :ok <- Client.check_redirect_uri(client, redirect_uri),
          true <- Client.grant_type_supported?(client, grant_type) do
       {:ok, client}
@@ -129,7 +129,7 @@ defmodule Boruta.Oauth.Authorization.Client do
         code_verifier: code_verifier
       )
       when not is_nil(id) and not is_nil(redirect_uri) do
-    with %Client{} = client <- ClientsAdapter.get_by(id: id),
+    with %Client{} = client <- ClientsAdapter.get_client(id),
          :ok <- Client.check_redirect_uri(client, redirect_uri),
          :ok <- validate_pkce(client, code_verifier),
          true <- Client.grant_type_supported?(client, grant_type) do